requirements.txt
gradio>=4.0.0
boto3>=1.28.0
botocore>=1.31.0
python-dotenv>=1.0.0
